// src/providerRuntime/types.ts
export type HttpRequest = {
    url: string;
    method: HttpRequestMethod;
    headers?: Record<string, string>;
    body?: Uint8Array;
};

/**
 * Possible errors that may happen during an HTTP request.
 */
export type HttpRequestError =
    | { type: "offline" }
    | { type: "no_route" }
    | { type: "connection_refused" }
    | { type: "timeout" }
    | { type: "response_too_big" }
    | { type: "server_error"; statusCode: number; response: Uint8Array }
    | { type: "other"; reason: string };

/**
 * HTTP request method.
 */
export type HttpRequestMethod =
    | "DELETE"
    | "GET"
    | "HEAD"
    | "POST";

/**
 * Response to an HTTP request.
 */
export type HttpResponse = {
    body: Uint8Array;
    headers: Record<string, string>;
    statusCode: number;
};
// ...
export type Result<T, E> =
    /**
     * Represents a successful result.
     */
    | { Ok: T }
    /**
     * Represents an error.
     */
    | { Err: E };